Init admin layout
