+++ client/components/streamer/StreamerMenu.vue
@@ -1,41 +1,9 @@
-<script lang="ts" setup>
-const active = ref(false);
-
-const route = useRoute();
-
-watch(
-  () => route.fullPath,
-  () => {
-    active.value = false;
-  }
-);
-</script>
+<script lang="ts" setup></script>
 
 <template>
-  <div class="mt-4">
-    <div
-      class="border rounded-lg border-border p-4 pt-8 sticky top-0 hidden lg:block"
-    >
-      <StreamerMenuContent />
-    </div>
-    <div class="lg:hidden">
-      <UButton @click="active = true" trailingIcon="i-heroicons-chevron-right">
-        Menu
-      </UButton>
-      <USlideover
-        v-model="active"
-        side="left"
-        :ui="{
-          background: 'bg-background dark:bg-background',
-          width: 'max-w-[320px]',
-        }"
-      >
-        <div class="p-4 py-6">
-          <StreamerMenuContent />
-        </div>
-      </USlideover>
-    </div>
-  </div>
+  <MenuContainer>
+    <StreamerMenuContent />
+  </MenuContainer>
 </template>
 
 <style scoped></style>
